+++ src/Queue/Jobs/NotificationRetryJob.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Glueful\Queue\Jobs;
 
+use Glueful\Bootstrap\ApplicationContext;
 use Glueful\Queue\Job;
 use Glueful\Tasks\NotificationRetryTask;
 use Glueful\Logging\LogManager;
@@ -46,9 +47,9 @@ use Throwable;
  */
 class NotificationRetryJob extends Job
 {
-    public function __construct(array $data = [])
+    public function __construct(array $data = [], ?ApplicationContext $context = null)
     {
-        parent::__construct($data);
+        parent::__construct($data, $context);
 
         // Set job configuration
         $this->queue = 'notifications';